requirements.txt
beautifulsoup4==4.4.1
click==6.3
cloudpickle==0.2.1
cssselect==0.9.1
cymem==1.31.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.10.1
Flask-SQLAlchemy==2.1
idna==2.0
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.8
joblib==0.9.4
lxml==3.5.0
MarkupSafe==0.23
murmurhash==0.26.1
newspaper3k==0.1.7
nltk==3.1
nose==1.3.7
numpy==1.10.4
Pillow==3.1.1
plac==0.9.1
preshed==0.46.2
psycopg2==2.6.1
python-dateutil==2.4.2
PyYAML==3.11
requests==2.9.1
scikit-learn==0.17.1
scipy==0.17.0
semver==2.4.0
six==1.10.0
spacy==0.100.5
sputnik==0.8.3
SQLAlchemy==1.0.12
text-unidecode==1.0
thinc==5.0.5
tldextract==1.7.5
toolz==0.7.4
ujson==1.35
Werkzeug==0.11.4
wheel==0.24.0
